Added basic error checking for Friend requests
Added error checking for the friend request view:
view_friend_request()
reject_friend_request()
accept_friend_request()
send_friend_request()

+++ mysite/users/views.py
@@ -1,4 +1,4 @@
-from django.http.response import HttpResponseRedirect
+from django.http.response import HttpResponseBadRequest, HttpResponseForbidden, HttpResponseRedirect
 from django.shortcuts import render, get_object_or_404
 from . import views
 from django.http import HttpResponse
@@ -70,10 +70,15 @@ def login_view(request):
 
 
 def send_friend_request(request, User_id):
+    if request.user.is_anonymous:
+        return HttpResponseForbidden("Please sign in")
     user = get_object_or_404(User,pk=User_id)
     print(user)
     request_profile = User_Profile.objects.get(user=request.user)
-    object_profile = User_Profile.objects.get(user_id=User_id)
+    #Checks if the object_profile is valid
+    object_profile = get_object_or_404(User_Profile, user_id=User_id)
+    #object_profile = User_Profile.objects.get(user_id=User_id)
+    #TODO: CHECK IF THE ACTOR IS ALREADY FOLLOWING THE OBJECT
     f_request, created = FriendRequest.objects.get_or_create(actor=request_profile, object=object_profile)
     print("Friend request created")
     print(f_request.summary())
@@ -81,9 +86,14 @@ def send_friend_request(request, User_id):
 
 def accept_friend_request(request, User_id):
     #User id is from the actor, the person who sent the friend request
+    #Error checking
+    if request.user.is_anonymous:
+        return HttpResponseForbidden("Please sign in")
     actor_user_profile = get_object_or_404(User_Profile, user_id=User_id)
     object_user_profile = get_object_or_404(User_Profile,user=request.user)
     f_request = FriendRequest.objects.filter(actor=actor_user_profile, object=object_user_profile)
+    if not f_request.exists():
+        return HttpResponseBadRequest("Friend request does not exist")
     #TODO: ADD THE ACTOR IN USER FOLLOWS
 
     actor_user = get_object_or_404(User,pk=User_id)
@@ -95,15 +105,21 @@ def accept_friend_request(request, User_id):
     return HttpResponseRedirect('/authors/requests/view-request/{}/'.format(request.user.id))
 
 def reject_friend_request(request, User_id):
+    if request.user.is_anonymous:
+        return HttpResponseForbidden("Please sign in")
     actor_user_profile = get_object_or_404(User_Profile, user_id=User_id)
     object_user_profile = get_object_or_404(User_Profile,user=request.user)
     f_request = FriendRequest.objects.filter(actor=actor_user_profile, object=object_user_profile)
-    #TO DO: CHECK IF THE FRIEND REQUEST EXISTS BEFORE DELETING
+    if not f_request.exists():  #Checks if the friend request exists
+        return HttpResponseBadRequest("Friend request does not exist")
     f_request.delete()
     print("{} deleted {}s' friend request".format(object_user_profile.displayName, actor_user_profile.displayName))
     return HttpResponseRedirect('/authors/requests/view-request/{}/'.format(request.user.id))
 
 def view_friend_requests(request, User_id):
+    #Makes sure that only the author can see the requests
+    if request.user.id != User_id:
+        return HttpResponseForbidden("You are forbidden")
     user_profile = get_object_or_404(User_Profile, user_id=User_id)
     recieved_requests = FriendRequest.objects.filter(object=user_profile)
     sent_requests = FriendRequest.objects.filter(actor=user_profile)
